requirements.txt
flask==2.3.2
mysql-connector-python==9.2.0
python-dotenv==1.0.0
gunicorn==23.0.0
pandas==2.3.0
sqlalchemy==2.0.41
flask_sqlalchemy
pymysql